Use uint16_t instead of uint8_t for storing matrix block size in engines

+++ engines/src/engine_nc_nonnewton_cpu.hpp
@@ -53,7 +53,7 @@ public:
   // IMPORTANT: all constants above have to be in agreement with acc_flux_op_set
 
   // number of variables per jacobian matrix block
-  const static uint8_t N_VARS_SQ = N_VARS * N_VARS;
+  const static uint16_t N_VARS_SQ = N_VARS * N_VARS;
 
   uint8_t get_n_vars() const { return N_VARS; };
   uint8_t get_n_ops() const { return N_OPS; };
